refactor: combine functional client tests with existing integration tests
